Use abstract test class pattern
