+++ src/main/java/com/powsybl/openloadflow/sa/DefaultContingencyActivePowerLossDistribution.java
@@ -59,7 +59,7 @@ public class DefaultContingencyActivePowerLossDistribution implements Contingenc
 
         if ((paramsOverride.isDistributedSlack(loadFlowParameters) || paramsOverride.isAreaInterchangeControl(openLoadFlowParameters)) && Math.abs(mismatch) > 0) {
             ActivePowerDistribution activePowerDistribution = ActivePowerDistribution.create(paramsOverride.getBalanceType(loadFlowParameters), openLoadFlowParameters.isLoadPowerFactorConstant(), openLoadFlowParameters.isUseActiveLimits());
-            var result = activePowerDistribution.run(network.getSynchronousNetworks().getFirst(), mismatch);
+            ActivePowerDistribution.Result result = activePowerDistribution.run(network.getSynchronousNetworks().getFirst(), mismatch);
             Reports.reportContingencyActivePowerLossDistribution(reportNode, mismatch * PerUnit.SB, result.remainingMismatch() * PerUnit.SB);
             return mismatch - result.remainingMismatch();
         }
